Отделил открытие попапа в отдельный файл
