+++ src/view/site-menu.js
@@ -1,5 +1,3 @@
-
-
 export const createSiteMenuTemplate = (films) => {
 
   let watchlistCount = 0;
